fix: orderIndex를 Double에서 BigDecimal로 수정

+++ src/main/java/tosiltosil/backend/common/domain/order/OrderManager.java
@@ -1,12 +1,13 @@
 package tosiltosil.backend.common.domain.order;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface OrderManager {
 
-    Double generateInitialOrderIndex();
+    BigDecimal generateInitialOrderIndex();
 
-    Double generateOrderIndexBetween(Double prevOrderIndex, Double nextOrderIndex);
+    BigDecimal generateOrderIndexBetween(BigDecimal prevOrderIndex, BigDecimal nextOrderIndex);
 
     <T extends Orderable> List<T> renewOrderIndexes(List<T> entities);
 }
