+++ src/main/java/tosiltosil/backend/module/goal/domain/request/GoalOrderChangeRequest.java
@@ -1,10 +1,12 @@
 package tosiltosil.backend.module.goal.domain.request;
 
+import java.math.BigDecimal;
+
 public record GoalOrderChangeRequest(
         Long goalId,
         Integer targetIndex,
-        Double prevOrderIndex,
-        Double nextOrderIndex
+        BigDecimal prevOrderIndex,
+        BigDecimal nextOrderIndex
 ) {
 
 }
