+++ src/main/java/tosiltosil/backend/module/goal/infrastructure/GoalJpaRepository.java
@@ -1,5 +1,6 @@
 package tosiltosil.backend.module.goal.infrastructure;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -13,5 +14,5 @@ public interface GoalJpaRepository extends JpaRepository<Goal, Long> {
     List<Goal> findByMemberIdAndCategoryIdOrderByOrderIndexAsc(UUID memberId, Long categoryId);
 
     @Query("SELECT MAX(g.orderIndex) FROM Goal g WHERE g.memberId = :memberId")
-    Optional<Double> findMaxOrderIndexByMemberId(@Param("memberId") UUID memberId);
+    Optional<BigDecimal> findMaxOrderIndexByMemberId(@Param("memberId") UUID memberId);
 }
